Zoltan2:  moving partitioning solution quality to evaluate partition
